+++ lib/kaminari/active_record_extension.rb
@@ -1,9 +1,8 @@
+require File.join(File.dirname(__FILE__), 'active_record_relation_methods')
 module Kaminari
-  DEFAULT_PER_PAGE = 25 unless defined? ::Kaminari::DEFAULT_PER_PAGE
-
   module ActiveRecordExtension
     extend ActiveSupport::Concern
-
+    include Kaminari::ConfigurationMethods
     included do
       def self.inherited(kls) #:nodoc:
         # TERRIBLE HORRIBLE NO GOOD VERY BAD HACK: inheritable_attributes is not yet set here on AR 3.0
@@ -20,39 +19,8 @@ module Kaminari
           scope :page, Proc.new {|num|
             limit(default_per_page).offset(default_per_page * ([num.to_i, 1].max - 1))
           } do
-            # Specify the <tt>per_page</tt> value for the preceding <tt>page</tt> scope
-            #   Model.page(3).per(10)
-            def per(num)
-              if (n = num.to_i) <= 0
-                self
-              else
-                limit(n).offset(offset_value / limit_value * n)
-              end
-            end
-
-            # Total number of pages
-            def num_pages
-              (except(:offset, :limit).count.to_f / limit_value).ceil
-            end
-
-            # Current page number
-            def current_page
-              (offset_value / limit_value) + 1
-            end
-          end
-
-          # Overrides the default per_page value per model
-          #   class Article < ActiveRecord::Base
-          #     paginates_per 10
-          #   end
-          def self.paginates_per(val)
-            @_default_per_page = val
-          end
-
-          # This model's default per_page value
-          # returns 25 unless explicitly overridden via <tt>paginates_per</tt>
-          def self.default_per_page
-            @_default_per_page || Kaminari::DEFAULT_PER_PAGE
+            include Kaminari::ActiveRecordRelationMethods
+            include Kaminari::PageScopeMethods
           end
         end
       end
