+++ lib/kaminari/active_record_relation_methods.rb
@@ -1,0 +1,10 @@
+module Kaminari
+  module ActiveRecordRelationMethods
+    extend ActiveSupport::Concern
+    module InstanceMethods
+      def pagination_count
+        except(:offset, :limit).count
+      end
+    end
+  end
+end
