+++ lib/kaminari/mongoid_criteria_methods.rb
@@ -1,0 +1,16 @@
+module Kaminari
+  module MongoidCriteriaMethods
+    extend ActiveSupport::Concern
+    module InstanceMethods
+      def limit_value
+        options[:limit]
+      end
+      def offset_value
+        options[:skip]
+      end
+      def pagination_count
+        count
+      end
+    end
+  end
+end
