+++ lib/kaminari/mongoid_extension.rb
@@ -1,17 +1,17 @@
+require File.join(File.dirname(__FILE__), 'mongoid_criteria_methods')
 module Kaminari
-  DEFAULT_PER_PAGE = 25 unless defined? ::Kaminari::DEFAULT_PER_PAGE
-
   module MongoidExtension
     module Criteria
       extend ActiveSupport::Concern
 
       included do
-        delegate :page, :per, :num_pages, :current_page, :limit_value, :to => '@klass'
+        delegate :page, :per, :num_pages, :current_page, :limit_value, :offset_value, :pagination_count, :to => '@klass'
       end
     end
 
     module Document
       extend ActiveSupport::Concern
+      include Kaminari::ConfigurationMethods
 
       included do
         # Fetch the values at the specified page number
@@ -19,43 +19,8 @@ module Kaminari
         scope :page, Proc.new {|num|
           limit(default_per_page).offset(default_per_page * ([num.to_i, 1].max - 1))
         } do
-          # Specify the <tt>per_page</tt> value for the preceding <tt>page</tt> scope
-          #   Model.page(3).per(10)
-          def per(num)
-            if (n = num.to_i) <= 0
-              self
-            else
-              limit(n).offset(options[:skip] / options[:limit] * n)
-            end
-          end
-
-          # Total number of pages
-          def num_pages
-            (count.to_f / options[:limit]).ceil
-          end
-
-          # Current page number
-          def current_page
-            (options[:skip] / options[:limit]) + 1
-          end
-
-          def limit_value
-            options[:limit]
-          end
-        end
-
-        # Overrides the default per_page value per model
-        #   class Article < ActiveRecord::Base
-        #     paginates_per 10
-        #   end
-        def self.paginates_per(val)
-          @_default_per_page = val
-        end
-
-        # This model's default per_page value
-        # returns 25 unless explicitly overridden via <tt>paginates_per</tt>
-        def self.default_per_page
-          @_default_per_page || Kaminari::DEFAULT_PER_PAGE
+          include Kaminari::MongoidCriteriaMethods
+          include Kaminari::PageScopeMethods
         end
       end
     end
