+++ lib/kaminari/railtie.rb
@@ -5,8 +5,11 @@ begin; require 'active_record'; rescue LoadError; end
 begin; require 'mongoid'; rescue LoadError; end
 
 require File.join(File.dirname(__FILE__), 'helpers')
+require File.join(File.dirname(__FILE__), 'page_scope_methods')
+require File.join(File.dirname(__FILE__), 'configuration_methods')
 
 module Kaminari
+  DEFAULT_PER_PAGE = 25 unless defined? ::Kaminari::DEFAULT_PER_PAGE
   class Railtie < ::Rails::Railtie #:nodoc:
     initializer 'kaminari' do |app|
       if defined? ::ActiveRecord
